+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add a setting to warp the pointer to the center of the screen after each frame
 * Implement SteamHTTP and SteamGameCoordinator
 * Users can use Proton instead of wine for launching Windows games
+* Catch memory shortage in RAM Search
 
 ### Changed